refactor() : product 관련 패키지, global 패키지 분리
