+++ src/main/java/gift/product/business/dto/ProductRegisterDto.java
@@ -1,6 +1,6 @@
-package gift.business.dto;
+package gift.product.business.dto;
 
-import gift.persistence.entity.Product;
+import gift.product.persistence.entity.Product;
 
 public record ProductRegisterDto(
     String name,
